+++ src/outputsize.jl
@@ -1,6 +1,7 @@
 module NilNumber
 
 using NNlib
+import Random
 
 """
     Nil <: Number
@@ -39,6 +40,8 @@ Base.typemin(::Type{Nil}) = nil
 Base.typemax(::Type{Nil}) = nil
 
 Base.promote_rule(x::Type{Nil}, y::Type{<:Number}) = Nil
+
+Random.rand(rng::Random.AbstractRNG, ::Random.SamplerType{Nil}) = nil
 
 end  # module